fix(remix-node): destroying an empty file session tries to delete the entire session directory

+++ packages/remix-node/sessions/fileStorage.ts
@@ -88,6 +88,11 @@ export function createFileSessionStorage({
       await fsp.writeFile(file, content, "utf-8");
     },
     async deleteData(id) {
+      // An empty id will cause the session directory to be deleted
+      // and will throw a "EPERM" error.
+      if (!id) {
+        return;
+      }
       try {
         await fsp.unlink(getFile(dir, id));
       } catch (error: any) {
